# platform/api/openapi.json
{
  "components": {
    "schemas": {
...
      "AlterarVisibilidadeRequest": {
        "properties": {
          "revelacao": {
            "$ref": "#/components/schemas/RevelacaoContrato"
          },
          "versao_esperada": {
            "minimum": 0.0,
            "title": "Versao Esperada",
            "type": "integer"
          },
          "visibilidade": {
            "enum": [
              "mesa",
              "narrador"
            ],
            "title": "Visibilidade",
            "type": "string"
          }
        },
        "required": [
          "visibilidade",
          "versao_esperada"
        ],
        "title": "AlterarVisibilidadeRequest",
        "type": "object"
      },
...
      "CriarEntidadeRequest": {
        "properties": {
          "ficha": {
            "$ref": "#/components/schemas/FichaContrato"
          },
          "proprietario_id": {
            "anyOf": [
              {
                "maxLength": 100,
                "minLength": 1,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Proprietario Id"
          },
          "revelacao": {
            "$ref": "#/components/schemas/RevelacaoContrato"
          },
          "tipo": {
            "enum": [
              "personagem",
              "npc",
              "monstro"
            ],
            "title": "Tipo",
            "type": "string"
          },
          "visibilidade": {
            "default": "narrador",
            "enum": [
              "mesa",
              "narrador"
            ],
            "title": "Visibilidade",
            "type": "string"
          }
        },
        "required": [
          "tipo",
          "ficha"
        ],
        "title": "CriarEntidadeRequest",
        "type": "object"
      },
...
      "EfeitoComandoResposta": {
        "properties": {
          "efeito": {
            "$ref": "#/components/schemas/EfeitoResumo"
          },
          "versao": {
            "title": "Versao",
            "type": "integer"
          }
        },
        "required": [
          "versao",
          "efeito"
        ],
        "title": "EfeitoComandoResposta",
        "type": "object"
      },
...
      "EntidadePublica": {
        "properties": {
          "id": {
            "title": "Id",
            "type": "string"
          },
          "imagem": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "description": "Retrato em base64, quando revelado.",
            "title": "Imagem"
          },
          "nome_publico": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Nome Publico"
          }
        },
        "required": [
          "id"
        ],
        "title": "EntidadePublica",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "PersonagemResumo": {
        "properties": {
          "excluido_em": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Excluido Em"
          },
          "id": {
            "title": "Id",
            "type": "string"
          },
          "mesa_id": {
            "title": "Mesa Id",
            "type": "string"
          },
          "nome": {
            "title": "Nome",
            "type": "string"
          },
          "proprietario_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Proprietario Id"
          },
          "restauravel_ate": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Restauravel Ate"
          },
          "revelacao": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/RevelacaoContrato"
              },
              {
                "type": "null"
              }
            ]
          },
          "tipo": {
            "enum": [
              "personagem",
              "npc",
              "monstro"
            ],
            "title": "Tipo",
            "type": "string"
          },
          "versao": {
            "title": "Versao",
            "type": "integer"
          },
          "visibilidade": {
            "enum": [
              "mesa",
              "narrador"
            ],
            "title": "Visibilidade",
            "type": "string"
          }
        },
        "required": [
          "id",
          "mesa_id",
          "nome",
          "tipo",
          "visibilidade",
          "proprietario_id",
          "versao"
        ],
        "title": "PersonagemResumo",
        "type": "object"
      },
...
      "TransicaoEfeitoRequest": {
        "properties": {
          "motivo": {
            "anyOf": [
              {
                "maxLength": 300,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Motivo"
          },
          "versao_esperada": {
            "minimum": 0.0,
            "title": "Versao Esperada",
            "type": "integer"
          }
        },
        "required": [
          "versao_esperada"
        ],
        "title": "TransicaoEfeitoRequest",
        "type": "object"
      },
...
    },
    "securitySchemes": {
      "HTTPBearer": {
        "scheme": "bearer",
        "type": "http"
      }
    }
  },
  "info": {
    "description": "API autoritativa da plataforma colaborativa Cursed.",
    "title": "Cursed — Plataforma Colaborativa",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/mesas/{mesa_id}/entidades": {
      "post": {
        "description": "Personagem, NPC ou monstro do Narrador; oculto por padrão.",
        "operationId": "criar_entidade_mesas__mesa_id__entidades_post",
        "parameters": [
          {
            "in": "path",
            "name": "mesa_id",
            "required": true,
            "schema": {
              "title": "Mesa Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CriarEntidadeRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PersonagemResumo"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Criar Entidade",
        "tags": [
          "Ferramentas do Narrador"
        ]
      }
    },
    "/mesas/{mesa_id}/entidades-publicas": {
      "get": {
        "description": "O que qualquer participante vê: nome público e retrato, nunca a ficha.",
        "operationId": "listar_entidades_publicas_mesas__mesa_id__entidades_publicas_get",
        "parameters": [
          {
            "in": "path",
            "name": "mesa_id",
            "required": true,
            "schema": {
              "title": "Mesa Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/EntidadePublica"
                  },
                  "title": "Response Listar Entidades Publicas Mesas  Mesa Id  Entidades Publicas Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Listar Entidades Publicas",
        "tags": [
          "Ferramentas do Narrador"
        ]
      }
    },
...
    "/mesas/{mesa_id}/personagens/{personagem_id}/efeitos": {
      "get": {
        "operationId": "listar_efeitos_mesas__mesa_id__personagens__personagem_id__efeitos_get",
        "parameters": [
          {
            "in": "path",
            "name": "mesa_id",
            "required": true,
            "schema": {
              "title": "Mesa Id",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "personagem_id",
            "required": true,
            "schema": {
              "title": "Personagem Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/EfeitoResumo"
                  },
                  "title": "Response Listar Efeitos Mesas  Mesa Id  Personagens  Personagem Id  Efeitos Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Listar Efeitos",
        "tags": [
          "Ficha viva"
        ]
      },
      "post": {
        "operationId": "aplicar_efeito_mesas__mesa_id__personagens__personagem_id__efeitos_post",
        "parameters": [
          {
            "in": "path",
            "name": "mesa_id",
            "required": true,
            "schema": {
              "title": "Mesa Id",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "personagem_id",
            "required": true,
            "schema": {
              "title": "Personagem Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AplicarEfeitoRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EfeitoComandoResposta"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Aplicar Efeito",
        "tags": [
          "Ferramentas do Narrador"
        ]
      }
    },
    "/mesas/{mesa_id}/personagens/{personagem_id}/efeitos/{efeito_id}": {
      "patch": {
        "operationId": "ajustar_efeito_mesas__mesa_id__personagens__personagem_id__efeitos__efeito_id__patch",
        "parameters": [
          {
            "in": "path",
            "name": "mesa_id",
            "required": true,
            "schema": {
              "title": "Mesa Id",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "personagem_id",
            "required": true,
            "schema": {
              "title": "Personagem Id",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "efeito_id",
            "required": true,
            "schema": {
              "title": "Efeito Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AjustarEfeitoRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EfeitoComandoResposta"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Ajustar Efeito",
        "tags": [
          "Ferramentas do Narrador"
        ]
      }
    },
    "/mesas/{mesa_id}/personagens/{personagem_id}/efeitos/{efeito_id}/{acao}": {
      "post": {
        "operationId": "transicionar_efeito_mesas__mesa_id__personagens__personagem_id__efeitos__efeito_id___acao__post",
        "parameters": [
          {
            "in": "path",
            "name": "mesa_id",
            "required": true,
            "schema": {
              "title": "Mesa Id",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "personagem_id",
            "required": true,
            "schema": {
              "title": "Personagem Id",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "efeito_id",
            "required": true,
            "schema": {
              "title": "Efeito Id",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "acao",
            "required": true,
            "schema": {
              "enum": [
                "suspender",
                "retomar",
                "encerrar"
              ],
              "title": "Acao",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TransicaoEfeitoRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EfeitoComandoResposta"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Transicionar Efeito",
        "tags": [
          "Ferramentas do Narrador"
        ]
      }
    },
...
    "/mesas/{mesa_id}/personagens/{personagem_id}/visibilidade": {
      "put": {
        "operationId": "alterar_visibilidade_mesas__mesa_id__personagens__personagem_id__visibilidade_put",
        "parameters": [
          {
            "in": "path",
            "name": "mesa_id",
            "required": true,
            "schema": {
              "title": "Mesa Id",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "personagem_id",
            "required": true,
            "schema": {
              "title": "Personagem Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AlterarVisibilidadeRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PersonagemResumo"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Alterar Visibilidade",
        "tags": [
          "Ferramentas do Narrador"
        ]
      }
    },
...
  }
}